Updated project endpoint file name and versions
